+++ Makefile
@@ -1,8 +1,49 @@
-all: run convert
+CXX     = g++
+CXXFLAGS = -Wall -Wextra
 
-run:
-	g++ -Dx_pos=5.0 -I include main.cpp -Ofast
-	./a > image.ppm     
+SRCS = $(shell find * -name "*.cpp")
+OBJS = $(SRCS:.cpp=.o)
+EXE  = rt
 
-convert:
-	magick image.ppm image.png
+DBGDIR = bin/debug
+DBGEXE = $(DBGDIR)/$(EXE)
+DBGOBJS = $(addprefix $(DBGDIR)/, $(OBJS))
+DBGCXXFLAGS = -g -O0 -DDEBUG
+
+RELDIR = bin/release
+RELEXE = $(RELDIR)/$(EXE)
+RELOBJS = $(addprefix $(RELDIR)/, $(OBJS))
+RELCXXFLAGS = -Ofast -DNDEBUG
+
+.PHONY: all clean debug prep release remake
+
+all: prep release
+
+debug: $(DBGEXE)
+
+$(DBGEXE): $(DBGOBJS)
+	$(CXX) $(CXXFLAGS) $(DBGCXXFLAGS) -o $(DBGEXE) $^
+
+$(DBGDIR)/%.o: %.cpp
+	$(CXX) -c $(CXXFLAGS) $(DBGCXXFLAGS) -o $@ $<
+
+release: $(RELEXE)
+
+$(RELEXE): $(RELOBJS)
+	$(CXX) $(CXXFLAGS) $(RELCXXFLAGS) -o $(RELEXE) $^
+
+$(RELDIR)/%.o: %.cpp
+	$(CXX) -c $(CXXFLAGS) $(RELCXXFLAGS) -o $@ $<
+
+prep:
+	@mkdir -p $(DBGDIR) $(RELDIR)
+	@mkdir -p $(DBGDIR)/geometry
+	@mkdir -p $(DBGDIR)/renderer
+	@mkdir -p $(RELDIR)/geometry
+	@mkdir -p $(RELDIR)/renderer
+
+remake: clean all
+
+clean:
+	rm -f $(RELEXE) $(RELOBJS) $(DBGEXE) $(DBGOBJS)
+
