+++ README.md
@@ -1,4 +1,4 @@
 # Ray_Tracing
 
 ### Last Rendered Image 🥳
-![alt text](https://github.com/binitkc220/Ray_Tracing/blob/master/image.png?raw=true)
+![alt text](https://github.com/binitkc220/Ray_Tracing/blob/bigdiff/asset/gen/image.png?raw=true)
