+++ geometry/hittable_list.h
@@ -7,7 +7,6 @@
 #include<vector>
 
 using std::shared_ptr;
-using std::make_shared;
 
 class hittable_list : public hittable {
     public:
